feat(credits): move per-user usage banner above the input bar
