+++ front/components/assistant/conversation/ContextUsageWarningBanner.tsx
@@ -1,11 +1,7 @@
 import { useCompactConversation } from "@app/hooks/conversations";
 import type { GetConversationContextUsageResponse } from "@app/lib/api/assistant/conversation/context_usage";
 import type { LightWorkspaceType } from "@app/types/user";
-import {
-  ContentMessageAction,
-  ContentMessageInline,
-  InfoCircle,
-} from "@dust-tt/sparkle";
+import { ContentMessageInline, Hoverable, InfoCircle } from "@dust-tt/sparkle";
 
 interface ContextUsageWarningBannerProps {
   owner: LightWorkspaceType;
@@ -27,24 +23,30 @@ export const ContextUsageWarningBanner = ({
     <ContentMessageInline
       icon={InfoCircle}
       variant="info"
-      className="mb-5 flex max-h-dvh w-full"
+      className="mb-2 flex w-full"
     >
-      <div className="flex min-w-0 items-center gap-2">
-        <span className="min-w-0 truncate text-foreground dark:text-foreground-night">
-          Conversation context is almost full.
+      <div className="flex w-full items-center justify-between gap-2">
+        <span className="min-w-0 truncate">
+          Conversation context is almost full
         </span>
+        {isCompacting ? (
+          <span className="copy-sm shrink-0 text-muted-foreground dark:text-muted-foreground-night">
+            Compacting
+          </span>
+        ) : (
+          <Hoverable
+            variant="primary"
+            className="copy-sm shrink-0 underline underline-offset-2"
+            onClick={() => {
+              if (contextUsage.model) {
+                void compact(contextUsage.model);
+              }
+            }}
+          >
+            Compact now
+          </Hoverable>
+        )}
       </div>
-      <ContentMessageAction
-        label={isCompacting ? "Compacting" : "Compact now"}
-        variant="outline"
-        size="xs"
-        disabled={isCompacting}
-        onClick={() => {
-          if (contextUsage.model) {
-            void compact(contextUsage.model);
-          }
-        }}
-      />
     </ContentMessageInline>
   );
 };
