+++ front/components/assistant/conversation/input_bar/InputBar.tsx
@@ -7,6 +7,7 @@ import InputBarContainer, {
   INPUT_BAR_ACTIONS,
 } from "@app/components/assistant/conversation/input_bar/InputBarContainer";
 import { InputBarContext } from "@app/components/assistant/conversation/input_bar/InputBarContext";
+import { InputBarUsageBanner } from "@app/components/assistant/conversation/input_bar/InputBarUsageBanner";
 import {
   INPUT_BAR_COMPACT_ENTER_ANIMATION_CLASSES,
   INPUT_BAR_COMPACT_MORPH_TRANSITION_CLASSES,
@@ -407,6 +408,7 @@ export const InputBar = React.memo(function InputBar({
         effectiveIsCompact && "min-w-0 flex-1"
       )}
     >
+      <InputBarUsageBanner owner={owner} />
       <PlanCard
         conversationId={conversation?.sId ?? null}
         workspaceId={owner.sId}
